// src/app/models/landFilter.ts
export class LandFilter {
    searchText: string = '';
    cityName: string = '';
    districtName: string = '';
    maxPrice: number | null = null;
    minPrice: number | null = null;
    minSquareMeter: number | null = null;
    maxSquareMeter: number | null = null;
    floorEquivalen: boolean = false;
    listingTypeName: string = ''
}